separated scraper into scraper.go
